# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ChatBot Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/conversations:
    get:
      operationId: fetchConversations
      tags: [chat]
      summary: Get all conversations for a user (one per bot)
      parameters:
        - in: query
          name: phone
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
  /chat/messages:
    get:
      operationId: fetchChatHistory
      tags: [chat]
      summary: Get chat history for a user with a bot
      parameters:
        - in: query
          name: botId
          required: true
          schema:
            type: string
        - in: query
          name: phone
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message to a bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: Bot reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "403":
          description: Access not unlocked
...
components:
  schemas:
...
    Conversation:
      type: object
      required: [botId, botName, botAvatar, messageCount, lastMessage, lastMessageAt]
      properties:
        botId:
          type: string
        botName:
          type: string
        botAvatar:
          type: string
        messageCount:
          type: integer
        lastMessage:
          type: string
        lastMessageAt:
          type: string
    ChatMessage:
      type: object
      required: [_id, botId, phone, role, text, createdAt]
      properties:
        _id:
          type: string
        botId:
          type: string
        phone:
          type: string
        role:
          type: string
        text:
          type: string
        createdAt:
          type: string
    ChatMessageInput:
      type: object
      required: [phone, botId, text]
      properties:
        phone:
          type: string
        botId:
          type: string
        text:
          type: string